# docs/tms_openapi.yml
openapi: 3.1.0
info:
  title: TMS portal server
  description: TMS portal server documentation
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://tms-auth-service.tacc.cloud
    description: Development Server

paths:
  "/login":
    post:
      tags:
        - Login
      summary: Login
      description: Login to TMS Portal using specified identity provider, and redirect back to redirect uri (must be |
        configured as an allowed redirect uri)
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TmsLoginRequest'
      responses:
        "307":
          description: Temporary redirect
          headers:
            location:
              schema:
                type: string
              description: redirect url
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'

  "/login/whoami":
    get:
      tags:
        - Login
      summary: whoami
      security:
        - Bearer: [ ]
      description: Get information about logged in user
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsWhoAmIResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'

  "/login/idps":
    get:
      tags:
        - Login
      summary: listIdps
      description: Get list of all supported login identity providers
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLoginIdpResponse'

  "/login/callback":
    get:
      tags:
        - Login
      summary: callback
      description: Called by identity provider with code that can be exchanged for a token.
      parameters:
        - name: code
          schema:
            type: string
          in: query
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsCallbackResponse'

  "/resources/{provider_id}/{provider_account_id}":
    get:
      tags:
        - Resource
      summary: listResources
      description: Get a list of the resources for a given provider
      security:
        - Bearer: [ ]
      parameters:
        - name: provider_id
          in: path
          required: true
          description: The id of the provider to list resource for
          schema:
            type: string
        - name: provider_account_id
          in: path
          required: true
          description: The account id of the user for this provider
          schema:
            type: string

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResourcesResponse'

  "/resources/providers":
    get:
      tags:
        - Resource
      summary: listResourceProviders
      description: Get a list of the resource providers
      security:
        - Bearer: [ ]
      parameters:
        - name: linked_only
          description: If set to true, only previously linked providers will be returned
          schema:
            type: bool
            default: false
          in: query
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResourceProvidersResponse'

  "/resources/providers/links":
    get:
      tags:
        - Resource
      summary: getLinkedResourceAccounts
      description: Get a list of the resource providers that have been linked to the current tms account
      security:
        - Bearer: [ ]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLinkedResourceProviderResponse'

  "/resources/providers/links/{resource_provider_link_id}":
    delete:
      tags:
        - Resource
      summary: unlinkResourceAccount
      description: Unlink a resource provider account from the current tms account
      security:
        - Bearer: [ ]
      parameters:
        - name: resource_provider_link_id
          description: The id of the resource provider link (as returned by getLinks)
          required: true
          schema:
            type: string
          in: path

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlinkResourceProviderResponse'

  "/resources/providers/authorize":
    get:
      tags:
        - Resource
      summary: authorizeWithResourceProvider
      description: Authorization for a resource provider
      security:
        - Bearer: [ ]
      parameters:
        - name: provider_id
          schema:
            type: string
          in: query
          required: true
        - name: redirect_url
          schema:
            type: string
          in: query
          required: true
        - name: state
          schema:
            type: string
          in: query
          required: false
      responses:
        "307":
          description: Temporary redirect
          headers:
            location:
              schema:
                type: string
              description: redirect url
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'

  "/oauth2/authorize":
    get:
      tags:
        - OAuth2
      summary: authorize
      description: authorize endpoint for oauth2
      parameters:
        - name: response_type
          description: Response type expected from authorize - currently only "code is" supported
          required: true
          schema:
            type: string
          in: query
        - name: response_type
          description: Response type expected from authorize - currently only "code is" supported
          required: true
          schema:
            type: string
          in: query
        - name: client_id
          description: Response type expected from authorize - currently only "code is" supported
          required: true
          schema:
            type: string
          in: query
        - name: redirect_uri
          description: Response type expected from authorize - currently only "code is" supported
          required: true
          schema:
            type: string
          in: query
        - name: scope
          description: Response type expected from authorize - currently only "code is" supported
          required: false
          schema:
            type: string
          in: query
        - name: state
          description: Response type expected from authorize - currently only "code is" supported
          required: false
          schema:
            type: string
          in: query
      responses:
        "307":
          description: Temporary redirect
          headers:
            location:
              schema:
                type: string
              description: redirect url


  "/oauth2/tokens":
    post:
      tags:
      - OAuth2
      summary: tokens
      description: OAuth2 tokens endpoint.  At present only the grant_type code is supported.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthTokenRequest'

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TmsErrorResponse'

# ------------------------------------------------------------------------------
# --- Components ---------------------------------------------------------------
# ------------------------------------------------------------------------------
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    # ------------------------------------------------------------------------------
    # --- Model objects ------------------------------------------------------------
    # ------------------------------------------------------------------------------
    TmsErrorResponse:
      type: object
      properties:
        status:
          type: string
        result:
          type: string

    TmsWhoAmIResponse:
      type: object
      properties:
        status:
          type: string
        result:
          properties:
            name:
              type: string
            user_name:
              type: string
            idpDisplayName:
              type: string
            organization:
              type: string
    
    
    TmsLoginRequest:
      required:
        - idp_id
        - redirect_uri
      type: object
      properties:
        idp_id:
          type: string
        redirect_uri:
          type: string

    ListLoginIdpResponse:
      type: object
      properties:
        status:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/IdentityProvider'

    ListResourcesResponse:
      type: object
      properties:
        status:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/Resource'

    ListResourceProvidersResponse:
      type: object
      properties:
        status:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/IdentityProvider'

    GetLinkedResourceProviderResponse:
      type: object
      properties:
        status:
          type: string
        result:
          type: array
          items:
            $ref: '#/components/schemas/ResourceProviderLink'

    UnlinkResourceProviderResponse:
      type: object
      properties:
        status:
          type: string
        result:
          $ref: '#/components/schemas/ResourceProviderLogins'

    TmsCallbackResponse:
      type: object
      properties:
        status:
          type: string
        result:
          properties:
            token:
              type: string

    IdentityProvider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        clientId:
          type: string
        oauth2TokenUrl:
          type: string
        userInfoUrl:
          type: string
        created:
          type: string
        updated:
          type: string

    ResourceProviderLink:
      type: object
      properties:
        tms_identity:
          type: string
        id:
          type: number
        resource_provider_id:
          type: string
        resource_provider_name:
          type: string
        resource_provider_uuid:
          type: string
        resource_provider_account:
          type: string
        last_login:
          type: string
        enabled:
          type: boolean

    ResourceProviderLogins:
      type: object
      properties:
        tms_identity:
          type: string
        id:
          type: number
        resource_provider_uuid:
          type: string
        resource_provider_account:
          type: string
        last_login:
          type: string
        enabled:
          type: boolean


    Resource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        provider_id:
          type: string
        provider_name:
          type: string

    OAuthTokenResponse:
      type: object
      properties:
        access_token:
          $ref: '#/components/schemas/AccessToken'

    AccessToken:
      type: object
      properties:
        access_token:
          type: string
        expires_at:
          type: string
        expires_in:
          type: i64,
        id_token:
          type: string
        jti:
          type: string

    OAuthTokenRequest:
      required:
        - client_id
        - client_secret
        - grant_type
        - redirect_uri
      type: object
      properties:
        client_id:
          type: string
        client_secret:
          type: string
        grant_type:
          type: string
          enum:
            - authorization_code
        redirect_uri:
          type: string,
        code:
          type: string
        refresh_token:
          type: string